feat: add new question to correspoding space
add_challenge_to_spaces is checking the new questions and marking them from the YAML file.
js_actions is adding the questions to the correspoding space.
